Store total counts on FetchingArray when fb returns a "count" attribute
Previously we were accessing the FetchingArray.count method, which was
returning the number of entries currently in the array, rather than the
total number of available entries. When displaying statistics, this
makes our numbers look small and sad.

When building a fetching array we now store a "count" attribute if
present in the total_count attribute.